fix: correctly compute monthly bills
AWS does not accept start and end date being the same, in the case of
previously used logic, the last day of last month would get added to the
exported costs, which, by logic, is incorrect.

now all dates are calculated D-1 from today's date, and if its the first
day of the month, it shows you the cost for the previous month instead.

Signed-off-by: Chinmay D. Pai <chinmaydpai@gmail.com>